requirements.txt
python-telegram-bot
requests
tabulate
structlog